changed meeting function

+++ resources/views/training/training_history.blade.php
@@ -51,9 +51,12 @@
                                                     <div class="dropdown">
                                                         <a href="javascript:void(0)" data-bs-toggle="dropdown" aria-expanded="false"><i class="fas fa-ellipsis-v align-middle ms-2 cursor-pointer"></i></a>
                                                         <div class="dropdown-menu dropdown-menu-animated">
-                                                            <a class="dropdown-item" href="{{route('training.show',$data->id)}}">View</a>  
-                                                            <a class="dropdown-item" href="{{route('training.edit',$data->id)}}">Edit</a>
-                                                            <a class="dropdown-item" href="javascript:void(0)" onclick="deleteItem('{{ route('training.destroy',$data->id) }}')">Delete</a>   
+                                                            <a class="dropdown-item" href="{{route('training.show',$data->id)}}">View</a>
+                                                            @if ($data->created_by == Auth::id() || auth()->user()->hasPermission('admin'))
+                                                                <a class="dropdown-item" href="{{route('training.edit',$data->id)}}">Edit</a>
+                                                                <a class="dropdown-item" href="{{route('training.attendance',encrypt($data->id))}}">Edit</a>
+                                                                <a class="dropdown-item" href="javascript:void(0)" onclick="deleteItem('{{ route('training.destroy',$data->id) }}')">Delete</a>
+                                                            @endif   
                                                         </div>
                                                     </div>
                                                 </td>
